+++ Senzan/SourceCode/Game/01_GameObject/00_MeshObject/00_Character/01_Player/State/Root/01_Action/01_Combat/02_AttackCombo_2/AttackCombo_2.cpp
@@ -154,7 +154,7 @@ void AttackCombo_2::LateUpdate()
 	Combat::LateUpdate();
 
 	// Œo‰ßŽžŠÔ‚ð‰ÁŽZ.
-	float actual_anim_speed = m_pOwner->m_AnimSpeed;
+	float actual_anim_speed = static_cast<float>(m_pOwner->m_AnimSpeed);
 	float delta_time = actual_anim_speed * m_pOwner->GetDelta();
 	m_currentTime += delta_time;
